Fix : Fixing folder, fixing after create send data, fixing bug on update jenis barang, fixing bug on bantuan
