+++ app/Http/Controllers/Api/Pengungsi/PengungsiController.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Controllers\Kebutuhan;
+namespace App\Http\Controllers\Api\Pengungsi;
 
 use App\Http\Controllers\Controller;
-use App\Models\Barang\Barang;
-use App\Models\Kebutuhan\Kebutuhan;
+use App\Models\Penduduk\Penduduk;
+use App\Models\Pengungsi\Pengungsi;
 use App\Models\Posko\Posko;
 use Carbon\Carbon;
 use Exception;
@@ -13,7 +13,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 
-class KebutuhanController extends Controller
+class PengungsiController extends Controller
 {
     public function __construct()
     {
@@ -29,35 +29,33 @@ class KebutuhanController extends Controller
     }
     public function index()
     {
-        $posko = Posko::whereHas('kebutuhan', function($query) {
-            $query->whereNotNull('IDPosko');
-        })->get();
-        return response()->json($posko, 200); // Mengembalikan response JSON dengan status 200
+        $pengungsi = Pengungsi::with(['penduduk', 'posko'])->get();
+        return response()->json($pengungsi, 200); // Mengembalikan response JSON dengan status 201
 
     }
 
-    public function show($id)  // id yang digunakan idposko
+    public function show($id)
     {
-        $pengungsi = Posko::where('IDPosko', $id)->with(['pengguna', 'kebutuhan'])->first();
+        $pengungsi = Pengungsi::where('IDPengungsi', $id)->with(['penduduk', 'posko'])->first();
         return response()->json($pengungsi, 200); // Mengembalikan response JSON dengan status 200
+
     }
 
-    public function create()
+    public function create(Request $request)
     {
-        $kebutuhan = Kebutuhan::groupBy('IDPosko')->pluck('IDPosko');
-        $product = Barang::get();
-        $posko = Posko::with('pengguna')->whereNotIn('IDPosko', $kebutuhan)->get();
-        $data['product'] = $product;
+        $penduduk = Penduduk::get();
+        $posko = Posko::with('pengguna')->get();
+        $data['penduduk'] = $penduduk;
         $data['posko'] = $posko;
         return response()->json($data, 200); // Mengembalikan response JSON dengan status 200
-
     }
 
     public function store(Request $request)
     {
         try {
             $validator = Validator::make($request->all(), [
-                'idPosko' => 'string|max:255', // Validasi problem
+                'idPosko' => 'required|numeric|max:50', // Validasi nomor kontak
+                'condition' => 'string|max:255', // Validasi problem
             ]);
 
             if ($validator->fails()) {
@@ -65,16 +63,14 @@ class KebutuhanController extends Controller
             }
 
             DB::beginTransaction();
-            foreach ($request->product as $product){
-                $kebutuhan = Kebutuhan::lockForUpdate()->create([
-                    'IDBarang' => $product['idProduct'],
-                    'IDPosko' => $request->idPosko,
-                    'JumlahKebutuhan' => $product['qty'],
-                    'LastUpdateDate' => Carbon::now(),
-                    'LastUpdateBy' => Auth::user()->id,
-                ]);
-            }
-            if ($kebutuhan) {
+            $pengungsi = Pengungsi::lockForUpdate()->create([
+                'IDPenduduk' => $request->idPenduduk,
+                'IDPosko' => $request->idPosko,
+                'KondisiKhusus' => $request->condition,
+                'LastUpdateDate' => Carbon::now(),
+                'LastUpdateBy' => Auth::user()->id,
+            ]);
+            if ($pengungsi) {
                 DB::commit();
                 return response()->json('data berhasil disimpan', 200); // Mengembalikan response JSON dengan status 201
             } else {
@@ -84,29 +80,44 @@ class KebutuhanController extends Controller
         } catch (Exception $e) {
             return response()->json(['error' => 'Data tidak dapat disimpan.'], 500); // Mengembalikan error jika terjadi pengecualian
             return response()->json(['error' => $e], 500); // Mengembalikan error jika terjadi pengecualian
+
         }
     }
 
-    public function qtyReceived(Request $request) // untuk mengisi jumlah yang diterima
+    public function edit($id)
+    {
+        $pengungsi = Pengungsi::where('IDPengungsi', $id)->with(['penduduk', 'posko'])->first();
+        $penduduk = Penduduk::get();
+        $posko = Posko::with('pengguna')->get();
+        $data['penduduk'] = $penduduk;
+        $data['posko'] = $posko;
+        $data['pengungsi'] = $pengungsi;
+        return response()->json($data, 200); // Mengembalikan response JSON dengan status 200
+
+    }
+
+    public function update(Request $request, $id)
     {
         try {
             $validator = Validator::make($request->all(), [
-                'idPosko' => 'string|max:255', // Validasi problem
+                'idPosko' => 'required|numeric|max:50', // Validasi nomor kontak
+                'condition' => 'string|max:255', // Validasi problem
             ]);
+
 
             if ($validator->fails()) {
                 return response()->json($validator->errors(), 422); // Gunakan status 422 untuk validasi gagal
             }
 
             DB::beginTransaction();
-            foreach ($request->product as $product){
-                $kebutuhan = Kebutuhan::where('IDPosko', $request->idPosko)->where('IDBarang', $product['idProduct'])->lockForUpdate()->update([
-                    'JumlahDiterima' => $product['qty'],
-                    'LastUpdateDate' => Carbon::now(),
-                    'LastUpdateBy' => Auth::user()->id,
-                ]);
-            }
-            if ($kebutuhan) {
+            $pengungsi = Pengungsi::where('IDPengungsi', $id)->lockForUpdate()->update([
+                'IDPenduduk' => $request->idPenduduk,
+                'IDPosko' => $request->idPosko,
+                'KondisiKhusus' => $request->condition,
+                'LastUpdateDate' => Carbon::now(),
+                'LastUpdateBy' => Auth::user()->id,
+            ]);
+            if ($pengungsi) {
                 DB::commit();
                 return response()->json('data berhasil disimpan', 200); // Mengembalikan response JSON dengan status 201
             } else {
@@ -114,10 +125,9 @@ class KebutuhanController extends Controller
                 return response()->json(['error' => 'Data posko tidak dapat disimpan.'], 500); // Mengembalikan error jika terjadi pengecualian
             }
         } catch (Exception $e) {
-            dd($e);
             return response()->json(['error' => 'Data tidak dapat disimpan.'], 500); // Mengembalikan error jika terjadi pengecualian
             return response()->json(['error' => $e], 500); // Mengembalikan error jika terjadi pengecualian
+
         }
     }
-
 }
